requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
huggingface-hub
numpy
requests